requirements.txt
streamlit
opencv-python
tensorflow
spotipy
numpy
Pillow
